# src/main/resources/openapi/openapi.swagger.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: GameStore API
  description: "A simple API to illustrate a hypothetical digital gamestore like GoG."
  contact:
    email: bakonyi.gergo.istvan@gmail.com

servers:
  - url: https://localhost:8000/gamestore

security:
  - apiKey: [ ]
tags:
  - name: Game
    description: Describe elements involved the products of the gamestore.
paths:
  /v1/store/games:
    post:
      description: Lets a user with admin level privilage to post a new game
      parameters:
        - $ref: "#/components/securitySchemes/accessToken"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Game"
      responses:
        "201":
          description: Returns the managed game detail view
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
        "400":
          description: asd
      security:
        - apiKey: [ ]
          accessToken: [ ]
    get:
      description: Return a list of games
      parameters:
        - name: title
          in: query
          description: "Search based on a game title"
          required: false
          schema:
            type: string
        - name: tags
          in: query
          description: "Search based on tags added by backoffice"
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: Either returns a list of games based on the provided query or an empty list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
      security:
        - apiKey: [ ]
  /v1/store/games/{id}:
    get:
      description: Returns a specific game, identified by its id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully returned a list of artists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
      security:
        - apiKey: [ ]

  /v1/store/games/{gameId}/detail:
    post:
      parameters:
        - $ref: "#/components/securitySchemes/accessToken"
        - name: gameId
          in: path
          required: true
          schema:
            type: string
        - name: detailId
          in: path
          required: true
          schema:
            type: string
      description: Adds a new game detail paage for the requested game. Fundamentally inactive status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameDetail"
      responses:
        "200":
          description: Returns the managed game detail view
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
      security:
        - apiKey: [ ]
          accessToken: [ ]
  /v1/store/games/{gameId}/detail/{detailId}:
    patch:
      parameters:
        - $ref: "#/components/securitySchemes/accessToken"
        - name: gameId
          in: path
          required: true
          schema:
            type: string
        - name: priceId
          in: path
          required: true
          schema:
            type: string
        - name: fields
          in: query
          required: true
          schema:
            type: string
      description: This endpoints allows the user to modify a given detail activity. Only allowed field now is the "active" flag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameDetail"
      responses:
        "200":
          description: Returns the managed game detail view
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
      security:
        - apiKey: [ ]
          accessToken: [ ]
  /v1/store/games/{id}/prices:
    post:
      parameters:
        - $ref: "#/components/securitySchemes/accessToken"
        - name: id
          in: path
          required: true
          schema:
            type: string
      description: Add a new draft price for the game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Price"
      responses:
        "200":
          description: Returns the new game price
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Price"
      security:
        - apiKey: [ ]
          accessToken: [ ]
  /v1/store/games/{gameId}/prices/{priceId}:
    patch:
      parameters:
        - $ref: "#/components/securitySchemes/accessToken"
        - name: gameId
          in: path
          required: true
          schema:
            type: string
        - name: priceId
          in: path
          required: true
          schema:
            type: string
        - name: fields
          in: query
          required: true
          schema:
            type: string
      description: With this endpoint the backoffice user could set the given price as the current, active price. The only allowed field is the "active" flag. If a price was mistakely given do not try to edit other parts of the document. Only the flag could be edited.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Price"
      responses:
        "200":
          description: Returns the managed price
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Price"
      security:
        - apiKey: [ ]
          accessToken: [ ]
  /v1/user:
    post:
      description: Registration of a user
      parameters:
        #- $ref: "#/components/securitySchemes/apiKey"
        - $ref: "#/components/securitySchemes/accessToken"
        - name: "body"
          in: "path"
          description: "body"
          required: true
          schema:
            $ref: "#/components/schemas/User"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserToken"
          description: A UserToken object containing a JWT accessToken if the "registration" is complete
        "400":
          description: If any error occures.
      security:
        - apiKey: [ ]
  /v1/user/{id}:
    get:
      description: Obtain public information about the requested users personal information and game library
      parameters:
        - $ref: "#/components/securitySchemes/apiKey"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully returned an artist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserLibrary"
      security:
        - apiKey: [ ]
  /v1/cart:
    post:
      parameters:
        #- $ref: "#/components/securitySchemes/apiKey"
        - $ref: "#/components/securitySchemes/accessToken"
      description: "With this operation a user can start a purchase"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShoppingCart"
      responses:
        "200":
          description: Successfully created a new cart item
        "400":
          description: asd
      security:
        - apiKey: [ ]
          accessToken: [ ]
    patch:
      description: With this operation a user can manage their shopping cart by replacing fields.
      parameters:
        - $ref: "#/components/securitySchemes/apiKey"
        # - $ref: "#/components/securitySchemes/accessToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShoppingCart"
      responses:
        "200":
          description: Successfully created a new cart item
      security:
        - apiKey: [ ]
          accessToken: [ ]
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-KEY
    accessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "A jwt token that contains user information"

  schemas:
    UserToken:
      type: object
      properties:
        accessToken:
          type: string
    User:
      type: object
      properties:
        username:
          type: string
    QueryTag:
      type: object
      properties:
        name:
          type: string
    Duration:
      type: object
      properties:
        begin:
          type: string
          format: date
        end:
          type: string
          format: date
    Price:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for every game entry.
        currency:
          type: string
        netPrice:
          type: number
        grossPrice:
          type: number
        region:
          type: string
        sale:
          type: string
        onSale:
          type: boolean
        active:
          type: boolean
    Game:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for every game entry.
        title:
          type: string
        description:
          type: string
        queryTag:
          type: array
          items:
            $ref: "#/components/schemas/QueryTag"
        releaseDate:
          type: string
          format: date
        prices:
          type: array
          items:
            $ref: "#/components/schemas/Price"
    ShoppingCart:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for every game entry.
        userId:
          type: string
        games:
          type: array
          items:
            $ref: "#/components/schemas/Game"
        total:
          $ref: "#/components/schemas/Price"
        state:
          type: string
          enum: [ "paid", "unpaid", "deleted" ]
    UserLibrary:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        games:
          type: array
          items:
            $ref: "#/components/schemas/Game"
    GameDetail:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for every game entry.
        title:
          type: string
        description:
          type: string
        queryTag:
          type: array
          items:
            $ref: "#/components/schemas/QueryTag"
        releaseDate:
          type: string
          format: date
        active:
          type: boolean